refac: model 및 serializer 리팩토링
