+++ webapp/admin.py
@@ -1,7 +1,8 @@
 from django.contrib import admin
-from .models import *
+from webapp.models import organization
+
 
 # Register your models here.
-@admin.register(sales.Organization)
+@admin.register(organization.Organization)
 class Organization(admin.ModelAdmin):
     list_display = ["id", "name"]
