+++ webapp/models/__init__.py
@@ -1,2 +1,2 @@
 # __init__.py
-from .sales import *
+from .sale import *
